DM-36723: fix conda Jenkins build

+++ python/lsst/ts/standardscripts/base_script_test_case.py
@@ -114,6 +114,7 @@ class BaseScriptTestCase(metaclass=abc.ABCMeta):
         ) as remote:
 
             initial_path = os.environ["PATH"]
+            process = None
             try:
                 os.environ["PATH"] = str(script_path.parent) + ":" + initial_path
                 process = await asyncio.create_subprocess_exec(
@@ -123,7 +124,8 @@ class BaseScriptTestCase(metaclass=abc.ABCMeta):
                 state = await remote.evt_state.next(flush=False, timeout=MAKE_TIMEOUT)
                 assert state.state == Script.ScriptState.UNCONFIGURED
             finally:
-                process.terminate()
+                if process is not None:
+                    process.terminate()
                 os.environ["PATH"] = initial_path
 
     async def configure_script(self, **kwargs):
